#129 WIP: factor out node editing logic
